[fix] Remove unnecesarry varibles about the number of grid
